Widen filter_index parameter to int64_t in CE/MG tally scoring
score_general_ce_nonanalog, score_general_ce_analog, and score_general_mg took the flat results_ filter index as an int parameter. Their callers compute it from FilterBinIter::index_ (now int64_t), so the value was narrowed back to 32 bits at the call boundary and could index the wrong or out-of-bounds bin for a regular Monte Carlo tally exceeding 2^31 filter bins. This path is not used by the random ray solver; widen the parameter so large continuous-energy and multigroup MC tallies are correct too. Each function uses filter_index only for its final tally.results_(filter_index, ...) write, so widening the parameter is sufficient.

+++ src/tallies/tally_scoring.cpp
@@ -578,7 +578,7 @@ double get_nuclide_xs(const Particle& p, int i_nuclide, int score_bin)
 //! collision estimator.
 
 void score_general_ce_nonanalog(Particle& p, int i_tally, int start_index,
-  int filter_index, double filter_weight, int i_nuclide, double atom_density,
+  int64_t filter_index, double filter_weight, int i_nuclide, double atom_density,
   double flux)
 {
   Tally& tally {*model::tallies[i_tally]};
@@ -1112,7 +1112,7 @@ void score_general_ce_nonanalog(Particle& p, int i_tally, int start_index,
 //! is not used for analog tallies.
 
 void score_general_ce_analog(Particle& p, int i_tally, int start_index,
-  int filter_index, double filter_weight, int i_nuclide, double atom_density,
+  int64_t filter_index, double filter_weight, int i_nuclide, double atom_density,
   double flux)
 {
   Tally& tally {*model::tallies[i_tally]};
@@ -1615,7 +1615,7 @@ void score_general_ce_analog(Particle& p, int i_tally, int start_index,
 //! argument is really just used for filter weights.
 
 void score_general_mg(Particle& p, int i_tally, int start_index,
-  int filter_index, double filter_weight, int i_nuclide, double atom_density,
+  int64_t filter_index, double filter_weight, int i_nuclide, double atom_density,
   double flux)
 {
   auto& tally {*model::tallies[i_tally]};
